Add Rdb_buf_writer to wrap rdb_netbuf_store_* on a static buffer and add more length validation. Also fix a circulr dependency between rdb_buff.h and ha_rocksdb.h by introducing rdb_global.h
Upstream commit ID : fb-mysql-5.6.35/f55c956b4d8266bd6e4c8deabc2f69df34b6a3d4
PS-6864 : Merge fb-prod201902

Summary:
[Refactoring] Add Rdb_buf_writer to wrap rdb_netbuf_store_* on a static buffer and add more length validation. I only did a first round of refactoring in rdb_datadic.cc and didn't do everything - so that I can address the feedback without changing too much stuff. Will take care of the rest in another diff.

Also fix a circulr dependency between rdb_buff.h and ha_rocksdb.h by introducing rdb_global.h.

Differential Revision: D13625287

fbshipit-source-id: 762180b6a3d
